chyrp-backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy

fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]

fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart